Update and rename Task3/Task3-3/main.cpp to Задача 3/Задача 3-3/main.cpp
